GitHub ci pipeline (#9)
* added github build id support

* added github logging support

* Update ci.yml

* manual ci.yml

* Create main.yml

* Move ci content to main.yaml

* Delete .github/workflows/ci.yml

* Rename main.yml to ci.yml

* Rename ci.yml to main.yml

* Delete .github/workflows/main.yml

* Create main.yml

* Update and rename main.yml to ci.yml

* testing

* fixed azure devops platform check

* proper test

* just unit tests to start with

* set +e

* docker only runs interactive locally
fixed docker unit tests for non local runs

* fixed unit tests for non local run
aligned log implementation

* tweaked install package file

* unit test tweaks

* debug output

* ssh host keys auto accepted
fixed file watch test

* unit test fixes

* ssh+vcs unit test tweaks

* skip vcs clone update

---------

+++ .pipeliner/core/utils/test/files.class.sh
@@ -315,6 +315,8 @@ UnitTest_Files_Get_E2E_Tests() {
 # Files watch
 
 UnitTest_Files_Watch_Directory_Written() {
+  if [ ! -x "$(command -v inotifywait)" ] && [ ! -x "$(command -v fswatch)" ]; then exit 255 ; fi #skip
+
   #Given
   local directory=tmp
   rm -rf $directory 2>&1 > /dev/null
